Added sqlalchemy support. Config sorted out.
SQLAlchemy is finally in. This needs some clear documntation
on how it has been implemented as this was a bitch to implement
and I might forget how I did it later.

The instance folder is used for loading relative configs now.
If a relative config path is passed to binder, then the instance folder
is searched. The instance folder is located in the root of the git
repo at instance/. Also development, testing and production modes
habe been added to the app.ceh